requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
python-multipart
deepgram-sdk==3.3.2
langchain
langgraph
langsmith
google-generativeai>=0.3.0
google-cloud-aiplatform
langchain-google-genai
mem0ai
pandas>=1.0.0
pytest-asyncio
langchain-community
chromadb
sentence-transformers
langgraph-checkpoint-sqlite
